server/requirements.txt
numpy==1.24.3
Pillow==10.2.0
Flask==3.0.3
tensorflow==2.13.0
keras==2.13.1
flask-cors==4.0.1